Simplified the process, set up a mongo collection
