+++ .env
@@ -1,0 +1,1 @@
+PYTHONPATH=./backend
