--- .vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "python.pythonPath": "/home/shayanriyaz/anaconda3/envs/BestGeese/bin/python"
-}
