--- Old-Tests/app/models/__init__.py
@@ -1,8 +1,0 @@
-import os
-from person import Person
-from user import User
-from student import Student
-from subject import Subject
-
-from app import app,db, login_manager
-
