+++ backend/database/db.py
@@ -1,0 +1,7 @@
+from flask_mongoengine import MongoEngine 
+
+db = MongoEngine()
+
+def initialize_db(app):
+    db.init_app(app)
+
