--- backend/requirements.txt
@@ -1,2 +1,0 @@
-Flask==1.1.2 
-pymongo==3.11.0
